+++ NEWS.md
@@ -1,7 +1,8 @@
 
-# rquery 0.2.0 2017/12/26
+# rquery 0.2.0 2017/12/30
 
  * Don't store DB reference.
+ * Prepare for wrapr 1.1.0 features.
   
 # rquery 0.1.0 2017/12/03